패키지 올바른 위치에 배치
